tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/035.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: innerHTML adding frames with JS in and moving scripts</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>
  <div></div>
  <script>log('inline script #1');
    document.getElementsByTagName('div')[1].innerHTML = '<iframe src="pages/helloworld.html"></iframe>';
    log('end script #1');
  </script>

  <script type="text/javascript">
  log( 'inline script #2' );
  var t = async_test()

  function test() {
    assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'inline script #2', 'frame/popup script']);
    t.done();
    /*, ['inline script #1', 'end script #1', 'frame/popup script', 'inline script #2'] */
  }
  onload = t.step_func(function(){
    try{
      document.body.appendChild(document.importNode( top.frames[0].document.getElementsByTagName('script')[0], true ));
    }catch(e){ log('ERROR - tested functionality not supported'); }
    setTimeout(t.step_func(test), 200);
  });
  </script>

</body></html>
